requirements.txt
av
accelerate
aiofiles
aiohttp
boto3
botocore
cloudpickle
config
diffusers
einops
fastapi
fvcore
gradio
gradio-client
imageio
matplotlib
numpy
omegaconf
onnxruntime
opencv-python
pandas
peft
pillow
pycocotools
python-multipart
pyyaml
psutil
redis
regex==2024.5.15
requests
safetensors
scikit-image
scipy
timm
tokenizers
torch
torchaudio
torchmetrics
torchvision
tqdm
transformers
triton
uvicorn
